% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte.R
\docType{data}
\name{analyte}
\alias{analyte}
\title{Analyte Data}
\format{
A tibble with the columns:
\describe{
\item{Analyte}{Name of the analyte.}
\item{Units}{Units of the analyte.}
\item{MRL}{Minimum reporting level.}
\item{GeneralMethod}{General method.}
}
}
\usage{
analyte
}
\description{
A tibble of the different analytes.
}
\keyword{datasets}
