% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytesample.R
\docType{data}
\name{analytesample}
\alias{analytesample}
\title{Analyte Value Data}
\format{
A tibble with the columns:
\describe{
\item{LabID}{Unique lab identifier for sample.}
\item{Biosite}{Site name.}
\item{DateAnalyteSample}{Sample date.}
}
}
\usage{
analytesample
}
\description{
A tibble of the sites, dates and lab ids for visit to take water samples.
}
\keyword{datasets}
