% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytevalue.R
\docType{data}
\name{analytevalue}
\alias{analytevalue}
\title{Analyte Value Data}
\format{
A tibble with the columns:
\describe{
\item{LabID}{Unique lab identifier for sample.}
\item{Analyte}{Analyte Name.}
\item{Value}{Analyte value.}
}
}
\usage{
analytevalue
}
\description{
A tibble of the analyte values.
}
\keyword{datasets}
