% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytevisit.R
\docType{data}
\name{analytevisit}
\alias{analytevisit}
\title{Analyte Visit Data}
\format{
A tibble with the columns:
\describe{
\item{LabID}{Unique lab identifier for sample.}
\item{Biosite}{Site name.}
\item{DateAnalyteVisit}{Visit date.}
}
}
\usage{
analytevisit
}
\description{
A tibble of the sites, dates and lab ids for visit to take water samples.
}
\keyword{datasets}
