% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthiccount.R
\docType{data}
\name{benthiccount}
\alias{benthiccount}
\title{Benthic Count Data}
\format{
A tibble with the columns:
\describe{
\item{Biosite}{Site name.}
\item{DateBenthicSample}{Sample date.}
\item{Order}{Order name.}
\item{Family}{Family name.}
\item{Count}{Number of individuals.}
}
}
\usage{
benthiccount
}
\description{
A tibble of the benthic invertebrate counts by family.
}
\keyword{datasets}
