% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthicsample.R
\docType{data}
\name{benthicsample}
\alias{benthicsample}
\title{Benthic Sample Data}
\format{
A tibble with the columns:
\describe{
\item{Biosite}{Site name.}
\item{DateBenthicSample}{Sample date.}
\item{Subsample}{Proportion of subsampled.}
}
}
\usage{
benthicsample
}
\description{
A tibble of the sites and dates for benthic sampling.
}
\keyword{datasets}
