% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosite.R
\docType{data}
\name{biosite}
\alias{biosite}
\title{Biosite Data}
\format{
A tibble with the columns:
\describe{
\item{BioSite}{Site name}
\item{Creek}{Creek name}
\item{geometry}{Site spatial coordinates including elevation.}
}
}
\usage{
biosite
}
\description{
A tibble of the different sites for analyte, periphyton and benthic invertebrate sampling.
}
\keyword{datasets}
