% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effish.R
\docType{data}
\name{effish}
\alias{effish}
\title{Electrofishing Fish}
\format{
A tibble with the columns:
\describe{
\item{EFSite}{Site names}
\item{DateEFVisit}{Date of visit}
\item{EFPass}{Electrofishing pass number}
\item{FishNumber}{Fish number}
\item{Species}{Species code}
\item{ForkLength}{Fish fork length (mm)}
\item{BodyWeigth}{Fish body weight (g)}
}
}
\usage{
effish
}
\description{
A tibble of the fish caught by electrofishing.
}
\keyword{datasets}
