% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsite.R
\docType{data}
\name{efsite}
\alias{efsite}
\title{Electrofishing Site Data}
\format{
A tibble with the columns:
\describe{
\item{EFSite}{Electrofishing site name}
\item{Creek}{Creek name}
\item{SiteLength}{Site length (m)}
\item{Elevation}{Site elevation (masl)}
\item{DominantSubstrate}{Dominant substrate}
\item{geometry}{Site upstream spatial coordinates.}
}
}
\usage{
efsite
}
\description{
A tibble of the different backpack electrofishing sites.
}
\keyword{datasets}
