% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efspecies.R
\docType{data}
\name{efspecies}
\alias{efspecies}
\title{Electrofishing Species}
\format{
A tibble with the columns:
\describe{
\item{Species}{Species code}
\item{CommonName}{Species common name}
}
}
\usage{
efspecies
}
\description{
A tibble of the fish species caught by electrofishing.
}
\keyword{datasets}
