% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efvisit.R
\docType{data}
\name{efvisit}
\alias{efvisit}
\title{Electrofishing Visit}
\format{
A tibble with the columns:
\describe{
\item{EFSite}{Site names}
\item{DateEFVisit}{Date of visit}
\item{Conductivity}{Conductivity (uS/cm)}
\item{EFSecPass1}{Seconds electrofishing on first pass}
\item{EFSecPass2}{Seconds electrofishing on second pass}
\item{EFSecPass3}{Seconds electrofishing on third pass}
}
}
\usage{
efvisit
}
\description{
A tibble of the visits to electrofishing sites.
}
\keyword{datasets}
