% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periphyton.R
\docType{data}
\name{periphyton}
\alias{periphyton}
\title{Periphyton Data}
\format{
A tibble with the columns:
\describe{
\item{BioSite}{Site name}
\item{DatePeriphyton}{Sample date}
\item{Series}{Series number}
\item{SampleID}{Unique sample identifier}
\item{Chla}{Chlorophyll-A concentration}
}
}
\usage{
periphyton
}
\description{
A tibble of the periphyton time series Chlorophyll-A concentrations.
}
\keyword{datasets}
