% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\docType{data}
\name{site}
\alias{site}
\title{Biosite Data}
\format{
A tibble with the columns:
\describe{
\item{Site}{Section name.}
\item{Creek}{Creek name.}
\item{geometry}{Site spatial coordinates including elevation.}
}
}
\usage{
site
}
\description{
A tibble of the different sites
}
\keyword{datasets}
