% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\docType{data}
\name{taxon}
\alias{taxon}
\title{Taxon Data}
\format{
A tibble with the columns:
\describe{
\item{Order}{Order name.}
\item{Family}{Family name.}
}
}
\usage{
taxon
}
\description{
A tibble of the different taxa
}
\keyword{datasets}
