% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Temperature Data}
\format{
A tibble with the columns:
\describe{
\item{TempSite}{Site name}
\item{DateTimeTemperature}{Date and time of the temperature.}
\item{Temperature}{Water temperature (C)}
}
}
\usage{
temperature
}
\description{
A tibble of the water temperature data.
}
\keyword{datasets}
