% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempsite.R
\docType{data}
\name{tempsite}
\alias{tempsite}
\title{Temperature Site Data}
\format{
A tibble with the columns:
\describe{
\item{TempSite}{Site name}
\item{Creek}{Creek name}
\item{geometry}{Site spatial coordinates including elevation.}
}
}
\usage{
tempsite
}
\description{
A tibble of the different temperature sites.
}
\keyword{datasets}
