% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosite.R
\docType{data}
\name{biosite}
\alias{biosite}
\title{Biosite Data}
\format{
A tibble with the columns:
\describe{
\item{BioSite}{Name of the site.}
\item{Waterbody}{Name of the waterbody.}
\item{geometry}{Location of the site including elevation.}
}
}
\usage{
biosite
}
\description{
A tibble of the different sites
}
\keyword{datasets}
